+++ package.json
@@ -30,18 +30,18 @@
   },
   "dependencies": {
     "execa": "^7.2.0",
-    "ufo": "^1.1.2"
+    "ufo": "^1.2.0"
   },
   "devDependencies": {
-    "@types/node": "^20.4.6",
+    "@types/node": "^20.5.0",
     "@vitest/coverage-v8": "^0.34.1",
     "changelogen": "^0.5.4",
     "corepack": "^0.19.0",
-    "eslint": "^8.46.0",
+    "eslint": "^8.47.0",
     "eslint-config-unjs": "^0.2.1",
     "jiti": "^1.19.1",
     "pathe": "^1.1.1",
-    "prettier": "^2.8.8",
+    "prettier": "^3.0.2",
     "typescript": "^5.1.6",
     "unbuild": "^1.2.1",
     "vitest": "^0.34.1"
