+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from "./types";
-export * from "./utils/detect-package-manager";
+export * from "./package-manager";
 export * from "./api";
