+++ test/fixtures/yarn-berry/.pnp.loader.mjs
@@ -81,7 +81,7 @@ async function copyPromise(destinationFs, destination, sourceFs, source, opts) {
     : await sourceFs.lstatPromise(normalizedSource);
   await destinationFs.mkdirpPromise(
     destinationFs.pathUtils.dirname(destination),
-    { utimes: [atime, mtime] }
+    { utimes: [atime, mtime] },
   );
   const updateTime =
     typeof destinationFs.lutimesPromise === `function`
@@ -95,13 +95,13 @@ async function copyPromise(destinationFs, destination, sourceFs, source, opts) {
     normalizedDestination,
     sourceFs,
     normalizedSource,
-    { ...opts, didParentExist: true }
+    { ...opts, didParentExist: true },
   );
   for (const operation of prelayout) await operation();
   await Promise.all(
     postlayout.map((operation) => {
       return operation();
-    })
+    }),
   );
 }
 async function copyImpl(
@@ -112,7 +112,7 @@ async function copyImpl(
   destination,
   sourceFs,
   source,
-  opts
+  opts,
 ) {
   var _a, _b;
   const destinationStat = opts.didParentExist
@@ -136,7 +136,7 @@ async function copyImpl(
           sourceFs,
           source,
           sourceStat,
-          opts
+          opts,
         );
       }
       break;
@@ -152,7 +152,7 @@ async function copyImpl(
           sourceFs,
           source,
           sourceStat,
-          opts
+          opts,
         );
       }
       break;
@@ -168,7 +168,7 @@ async function copyImpl(
           sourceFs,
           source,
           sourceStat,
-          opts
+          opts,
         );
       }
       break;
@@ -193,7 +193,7 @@ async function copyImpl(
     (destinationStat.mode & 511) !== (sourceStat.mode & 511)
   ) {
     postlayout.push(() =>
-      destinationFs.chmodPromise(destination, sourceStat.mode & 511)
+      destinationFs.chmodPromise(destination, sourceStat.mode & 511),
     );
     updated = true;
   }
@@ -216,7 +216,7 @@ async function copyFolder(
   sourceFs,
   source,
   sourceStat,
-  opts
+  opts,
 ) {
   if (destinationStat !== null && !destinationStat.isDirectory()) {
     if (opts.overwrite) {
@@ -257,7 +257,7 @@ async function copyFolder(
           destinationFs.pathUtils.join(destination, entry),
           sourceFs,
           sourceFs.pathUtils.join(source, entry),
-          nextOpts
+          nextOpts,
         )
       ) {
         updated = true;
@@ -274,9 +274,9 @@ async function copyFolder(
           destinationFs.pathUtils.join(destination, entry),
           sourceFs,
           sourceFs.pathUtils.join(source, entry),
-          nextOpts
+          nextOpts,
         );
-      })
+      }),
     );
     if (entriesUpdateStatus.some((status) => status)) {
       updated = true;
@@ -290,7 +290,7 @@ function makeLinkOperation(
   destination,
   source,
   sourceStat,
-  linkStrategy
+  linkStrategy,
 ) {
   return async () => {
     await opFs.linkPromise(source, destination);
@@ -305,7 +305,7 @@ function makeCloneLinkOperation(
   destination,
   source,
   sourceStat,
-  linkStrategy
+  linkStrategy,
 ) {
   const isCloneSupported = isCloneSupportedCache.get(opFs);
   if (typeof isCloneSupported === `undefined`) {
@@ -314,7 +314,7 @@ function makeCloneLinkOperation(
         await opFs.copyFilePromise(
           source,
           destination,
-          fs.constants.COPYFILE_FICLONE_FORCE
+          fs.constants.COPYFILE_FICLONE_FORCE,
         );
         isCloneSupportedCache.set(opFs, true);
       } catch (err) {
@@ -325,7 +325,7 @@ function makeCloneLinkOperation(
             destination,
             source,
             sourceStat,
-            linkStrategy
+            linkStrategy,
           )();
         } else {
           throw err;
@@ -338,7 +338,7 @@ function makeCloneLinkOperation(
         opFs.copyFilePromise(
           source,
           destination,
-          fs.constants.COPYFILE_FICLONE_FORCE
+          fs.constants.COPYFILE_FICLONE_FORCE,
         );
     } else {
       return makeLinkOperation(
@@ -346,7 +346,7 @@ function makeCloneLinkOperation(
         destination,
         source,
         sourceStat,
-        linkStrategy
+        linkStrategy,
       );
     }
   }
@@ -361,7 +361,7 @@ async function copyFile(
   sourceFs,
   source,
   sourceStat,
-  opts
+  opts,
 ) {
   var _a;
   if (destinationStat !== null) {
@@ -381,13 +381,13 @@ async function copyFile(
             destination,
             source,
             sourceStat,
-            linkStrategy
+            linkStrategy,
           )
         : async () =>
             destinationFs.copyFilePromise(
               source,
               destination,
-              fs.constants.COPYFILE_FICLONE
+              fs.constants.COPYFILE_FICLONE,
             )
       : linkStrategy !== null
       ? makeLinkOperation(
@@ -395,12 +395,12 @@ async function copyFile(
           destination,
           source,
           sourceStat,
-          linkStrategy
+          linkStrategy,
         )
       : async () =>
           destinationFs.writeFilePromise(
             destination,
-            await sourceFs.readFilePromise(source)
+            await sourceFs.readFilePromise(source),
           );
   prelayout.push(async () => op());
   return true;
@@ -415,7 +415,7 @@ async function copySymlink(
   sourceFs,
   source,
   sourceStat,
-  opts
+  opts,
 ) {
   if (destinationStat !== null) {
     if (opts.overwrite) {
@@ -429,9 +429,9 @@ async function copySymlink(
     await destinationFs.symlinkPromise(
       convertPath(
         destinationFs.pathUtils,
-        await sourceFs.readlinkPromise(source)
+        await sourceFs.readlinkPromise(source),
       ),
-      destination
+      destination,
     );
   });
   return true;
@@ -484,7 +484,7 @@ class FakeFS {
         await Promise.all(
           entries.map((entry) => {
             return this.removePromise(this.pathUtils.resolve(p, entry));
-          })
+          }),
         );
       }
       for (let t = 0; t <= maxRetries; t++) {
@@ -548,7 +548,7 @@ class FakeFS {
           await this.utimesPromise(subPath, utimes[0], utimes[1]);
         } else {
           const parentStat = await this.statPromise(
-            this.pathUtils.dirname(subPath)
+            this.pathUtils.dirname(subPath),
           );
           await this.utimesPromise(subPath, parentStat.atime, parentStat.mtime);
         }
@@ -596,7 +596,7 @@ class FakeFS {
       stableSort = false,
       stableTime = false,
       linkStrategy = null,
-    } = {}
+    } = {},
   ) {
     return await copyPromise(this, destination, baseFs, source, {
       overwrite,
@@ -615,7 +615,7 @@ class FakeFS {
         this.copySync(
           this.pathUtils.join(destination, entry),
           baseFs.pathUtils.join(source, entry),
-          { baseFs, overwrite }
+          { baseFs, overwrite },
         );
       }
     } else if (stat.isFile()) {
@@ -634,7 +634,7 @@ class FakeFS {
       throw new Error(
         `Unsupported file type (file: ${source}, mode: 0o${stat.mode
           .toString(8)
-          .padStart(6, `0`)})`
+          .padStart(6, `0`)})`,
       );
     }
     const mode = stat.mode & 511;
@@ -750,7 +750,7 @@ class FakeFS {
             await new Promise((resolve) => setTimeout(resolve, interval));
           } else {
             throw new Error(
-              `Couldn't acquire a lock in a reasonable time (via ${lockPath})`
+              `Couldn't acquire a lock in a reasonable time (via ${lockPath})`,
             );
           }
         } else {
@@ -790,14 +790,14 @@ class FakeFS {
     return await this.writeFilePromise(
       p,
       `${JSON.stringify(data, null, 2)}
-`
+`,
     );
   }
   writeJsonSync(p, data) {
     return this.writeFileSync(
       p,
       `${JSON.stringify(data, null, 2)}
-`
+`,
     );
   }
   async preserveTimePromise(p, cb) {
@@ -833,7 +833,7 @@ function getEndOfLine(content) {
     (nl) =>
       nl ===
       `\r
-`
+`,
   ).length;
   const lf = matches.length - crlf;
   return crlf > lf
@@ -870,7 +870,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(p),
         flags,
         mode,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -883,12 +883,12 @@ class NodeFS extends BasePortableFakeFS {
         this.realFs.opendir(
           npath.fromPortablePath(p),
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.opendir(
           npath.fromPortablePath(p),
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     }).then((dir) => {
@@ -924,7 +924,7 @@ class NodeFS extends BasePortableFakeFS {
           } else {
             resolve(bytesRead);
           }
-        }
+        },
       );
     });
   }
@@ -938,7 +938,7 @@ class NodeFS extends BasePortableFakeFS {
           fd,
           buffer,
           offset,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         return this.realFs.write(
@@ -947,7 +947,7 @@ class NodeFS extends BasePortableFakeFS {
           offset,
           length,
           position,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -980,7 +980,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.realpath(
         npath.fromPortablePath(p),
         {},
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     }).then((path) => {
       return npath.toPortablePath(path);
@@ -988,7 +988,7 @@ class NodeFS extends BasePortableFakeFS {
   }
   realpathSync(p) {
     return npath.toPortablePath(
-      this.realFs.realpathSync(npath.fromPortablePath(p), {})
+      this.realFs.realpathSync(npath.fromPortablePath(p), {}),
     );
   }
   async existsPromise(p) {
@@ -1004,7 +1004,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.access(
         npath.fromPortablePath(p),
         mode,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1017,12 +1017,12 @@ class NodeFS extends BasePortableFakeFS {
         this.realFs.stat(
           npath.fromPortablePath(p),
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.stat(
           npath.fromPortablePath(p),
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -1056,12 +1056,12 @@ class NodeFS extends BasePortableFakeFS {
         this.realFs.lstat(
           npath.fromPortablePath(p),
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.lstat(
           npath.fromPortablePath(p),
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -1086,7 +1086,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.chmod(
         npath.fromPortablePath(p),
         mask,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1107,7 +1107,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(p),
         uid,
         gid,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1119,14 +1119,14 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.rename(
         npath.fromPortablePath(oldP),
         npath.fromPortablePath(newP),
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
   renameSync(oldP, newP) {
     return this.realFs.renameSync(
       npath.fromPortablePath(oldP),
-      npath.fromPortablePath(newP)
+      npath.fromPortablePath(newP),
     );
   }
   async copyFilePromise(sourceP, destP, flags = 0) {
@@ -1135,7 +1135,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(sourceP),
         npath.fromPortablePath(destP),
         flags,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1143,7 +1143,7 @@ class NodeFS extends BasePortableFakeFS {
     return this.realFs.copyFileSync(
       npath.fromPortablePath(sourceP),
       npath.fromPortablePath(destP),
-      flags
+      flags,
     );
   }
   async appendFilePromise(p, content, opts) {
@@ -1155,13 +1155,13 @@ class NodeFS extends BasePortableFakeFS {
           fsNativePath,
           content,
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.appendFile(
           fsNativePath,
           content,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -1183,13 +1183,13 @@ class NodeFS extends BasePortableFakeFS {
           fsNativePath,
           content,
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.writeFile(
           fsNativePath,
           content,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -1206,7 +1206,7 @@ class NodeFS extends BasePortableFakeFS {
     return await new Promise((resolve, reject) => {
       this.realFs.unlink(
         npath.fromPortablePath(p),
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1219,7 +1219,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(p),
         atime,
         mtime,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1236,7 +1236,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(p),
         atime,
         mtime,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1251,7 +1251,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.mkdir(
         npath.fromPortablePath(p),
         opts,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1264,12 +1264,12 @@ class NodeFS extends BasePortableFakeFS {
         this.realFs.rmdir(
           npath.fromPortablePath(p),
           opts,
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.rmdir(
           npath.fromPortablePath(p),
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       }
     });
@@ -1282,14 +1282,14 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.link(
         npath.fromPortablePath(existingP),
         npath.fromPortablePath(newP),
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
   linkSync(existingP, newP) {
     return this.realFs.linkSync(
       npath.fromPortablePath(existingP),
-      npath.fromPortablePath(newP)
+      npath.fromPortablePath(newP),
     );
   }
   async symlinkPromise(target, p, type) {
@@ -1298,7 +1298,7 @@ class NodeFS extends BasePortableFakeFS {
         npath.fromPortablePath(target.replace(/\/+$/, ``)),
         npath.fromPortablePath(p),
         type,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1306,7 +1306,7 @@ class NodeFS extends BasePortableFakeFS {
     return this.realFs.symlinkSync(
       npath.fromPortablePath(target.replace(/\/+$/, ``)),
       npath.fromPortablePath(p),
-      type
+      type,
     );
   }
   async readFilePromise(p, encoding) {
@@ -1316,7 +1316,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.readFile(
         fsNativePath,
         encoding,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1330,12 +1330,12 @@ class NodeFS extends BasePortableFakeFS {
         this.realFs.readdir(
           npath.fromPortablePath(p),
           { withFileTypes: true },
-          this.makeCallback(resolve, reject)
+          this.makeCallback(resolve, reject),
         );
       } else {
         this.realFs.readdir(
           npath.fromPortablePath(p),
-          this.makeCallback((value) => resolve(value), reject)
+          this.makeCallback((value) => resolve(value), reject),
         );
       }
     });
@@ -1353,7 +1353,7 @@ class NodeFS extends BasePortableFakeFS {
     return await new Promise((resolve, reject) => {
       this.realFs.readlink(
         npath.fromPortablePath(p),
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     }).then((path) => {
       return npath.toPortablePath(path);
@@ -1361,7 +1361,7 @@ class NodeFS extends BasePortableFakeFS {
   }
   readlinkSync(p) {
     return npath.toPortablePath(
-      this.realFs.readlinkSync(npath.fromPortablePath(p))
+      this.realFs.readlinkSync(npath.fromPortablePath(p)),
     );
   }
   async truncatePromise(p, len) {
@@ -1369,7 +1369,7 @@ class NodeFS extends BasePortableFakeFS {
       this.realFs.truncate(
         npath.fromPortablePath(p),
         len,
-        this.makeCallback(resolve, reject)
+        this.makeCallback(resolve, reject),
       );
     });
   }
@@ -1423,7 +1423,7 @@ class ProxiedFS extends FakeFS {
   async opendirPromise(p, opts) {
     return Object.assign(
       await this.baseFs.opendirPromise(this.mapToBase(p), opts),
-      { path: p }
+      { path: p },
     );
   }
   opendirSync(p, opts) {
@@ -1446,7 +1446,7 @@ class ProxiedFS extends FakeFS {
         buffer,
         offset,
         length,
-        position
+        position,
       );
     }
   }
@@ -1466,18 +1466,18 @@ class ProxiedFS extends FakeFS {
   createReadStream(p, opts) {
     return this.baseFs.createReadStream(
       p !== null ? this.mapToBase(p) : p,
-      opts
+      opts,
     );
   }
   createWriteStream(p, opts) {
     return this.baseFs.createWriteStream(
       p !== null ? this.mapToBase(p) : p,
-      opts
+      opts,
     );
   }
   async realpathPromise(p) {
     return this.mapFromBase(
-      await this.baseFs.realpathPromise(this.mapToBase(p))
+      await this.baseFs.realpathPromise(this.mapToBase(p)),
     );
   }
   realpathSync(p) {
@@ -1540,7 +1540,7 @@ class ProxiedFS extends FakeFS {
   async renamePromise(oldP, newP) {
     return this.baseFs.renamePromise(
       this.mapToBase(oldP),
-      this.mapToBase(newP)
+      this.mapToBase(newP),
     );
   }
   renameSync(oldP, newP) {
@@ -1550,14 +1550,14 @@ class ProxiedFS extends FakeFS {
     return this.baseFs.copyFilePromise(
       this.mapToBase(sourceP),
       this.mapToBase(destP),
-      flags
+      flags,
     );
   }
   copyFileSync(sourceP, destP, flags = 0) {
     return this.baseFs.copyFileSync(
       this.mapToBase(sourceP),
       this.mapToBase(destP),
-      flags
+      flags,
     );
   }
   async appendFilePromise(p, content, opts) {
@@ -1599,13 +1599,13 @@ class ProxiedFS extends FakeFS {
   async linkPromise(existingP, newP) {
     return this.baseFs.linkPromise(
       this.mapToBase(existingP),
-      this.mapToBase(newP)
+      this.mapToBase(newP),
     );
   }
   linkSync(existingP, newP) {
     return this.baseFs.linkSync(
       this.mapToBase(existingP),
-      this.mapToBase(newP)
+      this.mapToBase(newP),
     );
   }
   async symlinkPromise(target, p, type) {
@@ -1613,11 +1613,11 @@ class ProxiedFS extends FakeFS {
     if (this.pathUtils.isAbsolute(target))
       return this.baseFs.symlinkPromise(this.mapToBase(target), mappedP, type);
     const mappedAbsoluteTarget = this.mapToBase(
-      this.pathUtils.join(this.pathUtils.dirname(p), target)
+      this.pathUtils.join(this.pathUtils.dirname(p), target),
     );
     const mappedTarget = this.baseFs.pathUtils.relative(
       this.baseFs.pathUtils.dirname(mappedP),
-      mappedAbsoluteTarget
+      mappedAbsoluteTarget,
     );
     return this.baseFs.symlinkPromise(mappedTarget, mappedP, type);
   }
@@ -1626,11 +1626,11 @@ class ProxiedFS extends FakeFS {
     if (this.pathUtils.isAbsolute(target))
       return this.baseFs.symlinkSync(this.mapToBase(target), mappedP, type);
     const mappedAbsoluteTarget = this.mapToBase(
-      this.pathUtils.join(this.pathUtils.dirname(p), target)
+      this.pathUtils.join(this.pathUtils.dirname(p), target),
     );
     const mappedTarget = this.baseFs.pathUtils.relative(
       this.baseFs.pathUtils.dirname(mappedP),
-      mappedAbsoluteTarget
+      mappedAbsoluteTarget,
     );
     return this.baseFs.symlinkSync(mappedTarget, mappedP, type);
   }
@@ -1656,7 +1656,7 @@ class ProxiedFS extends FakeFS {
   }
   async readlinkPromise(p) {
     return this.mapFromBase(
-      await this.baseFs.readlinkPromise(this.mapToBase(p))
+      await this.baseFs.readlinkPromise(this.mapToBase(p)),
     );
   }
   readlinkSync(p) {
@@ -1704,11 +1704,11 @@ class VirtualFS extends ProxiedFS {
   static makeVirtualPath(base, component, to) {
     if (ppath.basename(base) !== `__virtual__`)
       throw new Error(
-        `Assertion failed: Virtual folders must be named "__virtual__"`
+        `Assertion failed: Virtual folders must be named "__virtual__"`,
       );
     if (!ppath.basename(component).match(VALID_COMPONENT))
       throw new Error(
-        `Assertion failed: Virtual components must be ended by an hexadecimal hash`
+        `Assertion failed: Virtual components must be ended by an hexadecimal hash`,
       );
     const target = ppath.relative(ppath.dirname(base), to);
     const segments = target.split(`/`);
@@ -1719,7 +1719,7 @@ class VirtualFS extends ProxiedFS {
       base,
       component,
       String(depth),
-      ...finalSegments
+      ...finalSegments,
     );
     return fullVirtualPath;
   }
@@ -1759,7 +1759,7 @@ class VirtualFS extends ProxiedFS {
     if (p === ``) return p;
     if (this.pathUtils.isAbsolute(p)) return VirtualFS.resolveVirtual(p);
     const resolvedRoot = VirtualFS.resolveVirtual(
-      this.baseFs.resolve(PortablePath.dot)
+      this.baseFs.resolve(PortablePath.dot),
     );
     const resolvedP = VirtualFS.resolveVirtual(this.baseFs.resolve(p));
     return ppath.relative(resolvedRoot, resolvedP) || PortablePath.dot;
@@ -1781,7 +1781,7 @@ const WATCH_MODE_MESSAGE_USES_ARRAYS =
   major > 19 || (major === 19 && minor >= 2) || (major === 18 && minor >= 13);
 
 const builtinModules = new Set(
-  Module.builtinModules || Object.keys(process.binding(`natives`))
+  Module.builtinModules || Object.keys(process.binding(`natives`)),
 );
 const isBuiltinModule = (request) =>
   request.startsWith(`node:`) || builtinModules.has(request);
@@ -1895,7 +1895,7 @@ async function load$1(urlString, context, nextLoad) {
     ((_a = context.importAssertions) == null ? void 0 : _a.type) !== `json`
   ) {
     const err = new TypeError(
-      `[ERR_IMPORT_ASSERTION_TYPE_MISSING]: Module "${urlString}" needs an import assertion of type "json"`
+      `[ERR_IMPORT_ASSERTION_TYPE_MISSING]: Module "${urlString}" needs an import assertion of type "json"`,
     );
     err.code = `ERR_IMPORT_ASSERTION_TYPE_MISSING`;
     throw err;
@@ -1903,8 +1903,8 @@ async function load$1(urlString, context, nextLoad) {
   if (process.env.WATCH_REPORT_DEPENDENCIES && process.send) {
     const pathToSend = pathToFileURL(
       npath.fromPortablePath(
-        VirtualFS.resolveVirtual(npath.toPortablePath(filePath))
-      )
+        VirtualFS.resolveVirtual(npath.toPortablePath(filePath)),
+      ),
     ).href;
     process.send({
       "watch:import": WATCH_MODE_MESSAGE_USES_ARRAYS
@@ -1961,7 +1961,7 @@ const ERR_PACKAGE_IMPORT_NOT_DEFINED = createErrorType(
       packagePath ? ` in package ${packagePath}package.json` : ``
     } imported from ${base}`;
   },
-  TypeError
+  TypeError,
 );
 const ERR_INVALID_MODULE_SPECIFIER = createErrorType(
   `ERR_INVALID_MODULE_SPECIFIER`,
@@ -1970,7 +1970,7 @@ const ERR_INVALID_MODULE_SPECIFIER = createErrorType(
       base ? ` imported from ${base}` : ``
     }`;
   },
-  TypeError
+  TypeError,
 );
 const ERR_INVALID_PACKAGE_TARGET = createErrorType(
   `ERR_INVALID_PACKAGE_TARGET`,
@@ -1983,7 +1983,7 @@ const ERR_INVALID_PACKAGE_TARGET = createErrorType(
     if (key === `.`) {
       assert(isImport === false);
       return `Invalid "exports" main target ${JSONStringify(
-        target
+        target,
       )} defined in the package config ${pkgPath}package.json${
         base ? ` imported from ${base}` : ``
       }${relError ? `; targets must start with "./"` : ``}`;
@@ -1991,12 +1991,12 @@ const ERR_INVALID_PACKAGE_TARGET = createErrorType(
     return `Invalid "${
       isImport ? `imports` : `exports`
     }" target ${JSONStringify(
-      target
+      target,
     )} defined for '${key}' in the package config ${pkgPath}package.json${
       base ? ` imported from ${base}` : ``
     }${relError ? `; targets must start with "./"` : ``}`;
   },
-  Error
+  Error,
 );
 const ERR_INVALID_PACKAGE_CONFIG = createErrorType(
   `ERR_INVALID_PACKAGE_CONFIG`,
@@ -2005,7 +2005,7 @@ const ERR_INVALID_PACKAGE_CONFIG = createErrorType(
       base ? ` while importing ${base}` : ``
     }${message ? `. ${message}` : ``}`;
   },
-  Error
+  Error,
 );
 
 function filterOwnProperties(source, keys) {
@@ -2046,7 +2046,7 @@ function getPackageConfig(path, specifier, base, readFileSyncFn) {
     throw new ERR_INVALID_PACKAGE_CONFIG(
       path,
       (base ? `"${specifier}" from ` : "") + fileURLToPath(base || specifier),
-      error.message
+      error.message,
     );
   }
   let { imports, main, name, type } = filterOwnProperties(packageJSON, [
@@ -2095,7 +2095,7 @@ function getPackageScopeConfig(resolved, readFileSyncFn) {
       fileURLToPath(packageJSONUrl),
       resolved,
       void 0,
-      readFileSyncFn
+      readFileSyncFn,
     );
     if (packageConfig2.exists) {
       return packageConfig2;
@@ -2146,7 +2146,7 @@ function throwImportNotDefined(specifier, packageJSONUrl, base) {
   throw new ERR_PACKAGE_IMPORT_NOT_DEFINED(
     specifier,
     packageJSONUrl && fileURLToPath(new URL(".", packageJSONUrl)),
-    fileURLToPath(base)
+    fileURLToPath(base),
   );
 }
 function throwInvalidSubpath(subpath, packageJSONUrl, internal, base) {
@@ -2156,7 +2156,7 @@ function throwInvalidSubpath(subpath, packageJSONUrl, internal, base) {
   throw new ERR_INVALID_MODULE_SPECIFIER(
     subpath,
     reason,
-    base && fileURLToPath(base)
+    base && fileURLToPath(base),
   );
 }
 function throwInvalidPackageTarget(
@@ -2164,7 +2164,7 @@ function throwInvalidPackageTarget(
   target,
   packageJSONUrl,
   internal,
-  base
+  base,
 ) {
   if (typeof target === "object" && target !== null) {
     target = JSONStringify(target, null, "");
@@ -2176,7 +2176,7 @@ function throwInvalidPackageTarget(
     subpath,
     target,
     internal,
-    base && fileURLToPath(base)
+    base && fileURLToPath(base),
   );
 }
 const invalidSegmentRegEx =
@@ -2190,7 +2190,7 @@ function resolvePackageTargetString(
   base,
   pattern,
   internal,
-  conditions
+  conditions,
 ) {
   if (subpath !== "" && !pattern && target[target.length - 1] !== "/")
     throwInvalidPackageTarget(match, target, packageJSONUrl, internal, base);
@@ -2217,7 +2217,7 @@ function resolvePackageTargetString(
   if (
     RegExpPrototypeExec(
       invalidSegmentRegEx,
-      StringPrototypeSlice(target, 2)
+      StringPrototypeSlice(target, 2),
     ) !== null
   )
     throwInvalidPackageTarget(match, target, packageJSONUrl, internal, base);
@@ -2235,7 +2235,7 @@ function resolvePackageTargetString(
   }
   if (pattern) {
     return new URL(
-      RegExpPrototypeSymbolReplace(patternRegEx, resolved.href, () => subpath)
+      RegExpPrototypeSymbolReplace(patternRegEx, resolved.href, () => subpath),
     );
   }
   return new URL(subpath, resolved);
@@ -2253,7 +2253,7 @@ function resolvePackageTarget(
   base,
   pattern,
   internal,
-  conditions
+  conditions,
 ) {
   if (typeof target === "string") {
     return resolvePackageTargetString(
@@ -2263,7 +2263,7 @@ function resolvePackageTarget(
       packageJSONUrl,
       base,
       pattern,
-      internal
+      internal,
     );
   } else if (ArrayIsArray(target)) {
     if (target.length === 0) {
@@ -2282,7 +2282,7 @@ function resolvePackageTarget(
           base,
           pattern,
           internal,
-          conditions
+          conditions,
         );
       } catch (e) {
         lastException = e;
@@ -2311,7 +2311,7 @@ function resolvePackageTarget(
         throw new ERR_INVALID_PACKAGE_CONFIG(
           fileURLToPath(packageJSONUrl),
           base,
-          '"exports" cannot contain numeric property keys.'
+          '"exports" cannot contain numeric property keys.',
         );
       }
     }
@@ -2327,7 +2327,7 @@ function resolvePackageTarget(
           base,
           pattern,
           internal,
-          conditions
+          conditions,
         );
         if (resolveResult === void 0) continue;
         return resolveResult;
@@ -2342,7 +2342,7 @@ function resolvePackageTarget(
     target,
     packageJSONUrl,
     internal,
-    base
+    base,
   );
 }
 function patternKeyCompare(a, b) {
@@ -2385,7 +2385,7 @@ function packageImportsResolve({ name, base, conditions, readFileSyncFn }) {
           base,
           false,
           true,
-          conditions
+          conditions,
         );
         if (resolveResult != null) {
           return resolveResult;
@@ -2401,7 +2401,7 @@ function packageImportsResolve({ name, base, conditions, readFileSyncFn }) {
             patternIndex !== -1 &&
             StringPrototypeStartsWith(
               name,
-              StringPrototypeSlice(key, 0, patternIndex)
+              StringPrototypeSlice(key, 0, patternIndex),
             )
           ) {
             const patternTrailer = StringPrototypeSlice(key, patternIndex + 1);
@@ -2415,7 +2415,7 @@ function packageImportsResolve({ name, base, conditions, readFileSyncFn }) {
               bestMatchSubpath = StringPrototypeSlice(
                 name,
                 patternIndex,
-                name.length - patternTrailer.length
+                name.length - patternTrailer.length,
               );
             }
           }
@@ -2430,7 +2430,7 @@ function packageImportsResolve({ name, base, conditions, readFileSyncFn }) {
             base,
             true,
             true,
-            conditions
+            conditions,
           );
           if (resolveResult != null) {
             return resolveResult;
@@ -2465,7 +2465,7 @@ async function resolvePrivateRequest(specifier, issuer, context, nextResolve) {
   } else {
     if (resolved.startsWith(`#`))
       throw new Error(
-        `Mapping from one private import to another isn't allowed`
+        `Mapping from one private import to another isn't allowed`,
       );
     return resolve$1(resolved, context, nextResolve);
   }
@@ -2478,7 +2478,7 @@ async function resolve$1(originalSpecifier, context, nextResolve) {
   let specifier = originalSpecifier;
   const url = tryParseURL(
     specifier,
-    isRelativeRegexp.test(specifier) ? context.parentURL : void 0
+    isRelativeRegexp.test(specifier) ? context.parentURL : void 0,
   );
   if (url) {
     if (url.protocol !== `file:`)
@@ -2499,7 +2499,7 @@ async function resolve$1(originalSpecifier, context, nextResolve) {
     if (subPath === `` && dependencyName !== `pnpapi`) {
       const resolved = pnpapi.resolveToUnqualified(
         `${dependencyName}/package.json`,
-        issuer
+        issuer,
       );
       if (resolved) {
         const content = await tryReadFile$1(resolved);
